# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple example NodeJS API project to demonstrate Swagger Documentation",
        "version": "1.0.0",
        "title": "Tasks API",
        "contact": {
            "email": "abc@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "schemes": ["http"],
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/orders": {
            "get": {
                "summary": "Get all orders",
                "produces": ["application/json"],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "date": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string"
                                    },
                                    "customer": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create new order",
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "order object",
                    "required": true,
                    "schema": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "string"
                            },
                            "title": {
                                "type": "string"
                            },
                            "date": {
                                "type": "string"
                            },
                            "type": {
                                "type": "string"
                            },
                            "customer": {
                                "type": "string"
                            }
                        }
                    }
                }],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "result": {
                                    "type": "object",
                                    "properties": {
                                        "acknowledged": { "type": "boolean" },
                                        "id": { "type": "string" }
                                    }
                                }
                            }
                        }
                    }
                }
            }

        },
        "/orders/{id}": {
            "get": {
                "summary": "Get a specific order by id",
                "produces": ["application/json"],
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "title": {
                                    "type": "string"
                                },
                                "date": {
                                    "type": "string"
                                },
                                "type": {
                                    "type": "string"
                                },
                                "customer": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/orders/{type}/{date}": {
            "get": {
                "summary": "Get orders by type and date",
                "produces": ["application/json"],
                "parameters": [{
                        "name": "type",
                        "in": "path",
                        "description": "Order type",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "date",
                        "in": "path",
                        "description": "Order date",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "type": {
                                    "type": "string"
                                },
                                "count": {
                                    "type": "number"
                                },
                                "orders": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                },
                                "related_customers": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}